feat: Refactored home chat and implemented list chats screen on mobile
